+++ example/components/DefaultDialog.tsx
@@ -6,7 +6,7 @@ import {
   DialogHeader,
   DialogProps,
   DialogTitle,
-} from "@ontech7/expo-dialog";
+} from "@ontech7/react-native-dialog";
 import { useCallback, useState } from "react";
 import { ColorValue, Text, TouchableOpacity, ViewStyle } from "react-native";
 import { ExclamationTriangleIcon } from "react-native-heroicons/outline";