+++ example/components/InputDialog.tsx
@@ -8,7 +8,7 @@ import {
   DialogInput,
   DialogProps,
   DialogTitle,
-} from "@ontech7/expo-dialog";
+} from "@ontech7/react-native-dialog";
 import { useCallback, useState } from "react";
 import {
   Alert,