+++ src/DialogProvider.tsx
@@ -1,7 +1,7 @@
-import { PortalProvider } from "@gorhom/portal";
 import React, { createContext, useContext } from "react";
 
 import type { TextStyle, ViewStyle } from "react-native";
+import { PortalProvider } from "./components/PortalProvider";
 
 export type DialogStyleConfig = {
   container?: ViewStyle;
@@ -32,12 +32,17 @@ export const useDialogStyles = () => useContext(DialogStyleContext);
 
 export type DialogProviderProps = {
   children: React.ReactNode;
-  styles?: DialogStyleConfig;
+  customStyles?: DialogStyleConfig;
 };
 
-export function DialogProvider({ children, styles }: DialogProviderProps) {
+export function DialogProvider({
+  children,
+  customStyles,
+}: DialogProviderProps) {
   return (
-    <DialogStyleContext.Provider value={{ ...defaultDialogStyles, ...styles }}>
+    <DialogStyleContext.Provider
+      value={{ ...defaultDialogStyles, ...customStyles }}
+    >
       <PortalProvider>{children}</PortalProvider>
     </DialogStyleContext.Provider>
   );