+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v1.1.0
+
+- Added `onPressOut` event on `Dialog` component
+- Added `animation` prop on `Dialog` component
+- Removed blur library (the user can import whatever blur library), no Expo necessary
+- Changed portal library
+- Removed `react-native-reanimated` and used `react-native` library for animations
+- Moved from `dist` to `build`
+
 ## v1.0.0
 
 - Add `Dialog` components